refactor: move metadata generation files to folder
